Update to support Python 3.4, and 3.5 (and later)
I don't have a test environment to test this unfortunately, so mostly I'm submitting this based upon my limited Python knowledge so that it can be fixed and tested.

I didn't want to simply update it to support only 3.5, since that'll break upgrades for anybody on 3.4. I also wanted to ensure that in future it can be easily extended to support 3.6 and onwards.